+++ tests/PaymentTest.php
@@ -10,19 +10,9 @@ class PaymentTest extends ClientTestCase
 {
     public function testResultObjectSerialization()
     {
-        $payment = new Payment\Payment();
-        $payment->setId('foo');
-        $payment->setType('bar');
-        $payment->setCurrency(null);
-        $payment->setAmount('55.00');
-        $payment->setPaymentMethod(Payment\Create::CARD);
+        $payment = $this->getPayment();
 
-        $card = new Payment\PaymentInstrumentCard();
-        $card->setCardBrand('Visa');
-        $card->setPan('4447');
-        $card->setExpYear(2021);
-        $card->setExpMonth(5);
-        $card->setHolder('John Smith');
+        $card = $this->getCard();
         $payment->setPaymentInstrument($card);
 
         $info = new Payment\AuthorizationInformation();
@@ -31,14 +21,14 @@ class PaymentTest extends ClientTestCase
         $payment->setAuthorizationInformation($info);
 
         $this->assertSame(
-            '{"id":"foo","amount":"55.00","type":"bar","payment_method":"card","payment_instrument":{"card_brand":"Visa","pan":"4447","exp_year":2024,"exp_month":5,"holder":"John Smith"},"authorization_information":{"url":"http:\/\/...","data":"some_data"}}',
+            '{"id":"foo","amount":"55.00","type":"bar","payment_method":"card","payment_instrument":{"card_brand":"Visa","pan":"4447","exp_year":'.(date('Y')+4).',"exp_month":11,"holder":"James Bond"},"authorization_information":{"url":"http:\/\/...","data":"some_data"}}',
             $payment->serialize()
         );
     }
 
     public function testResultObjectUnserialization()
     {
-        $json = '{"id":"foo","amount":"55.00","type":"bar","payment_method":"card","payment_instrument":{"card_brand":"Visa","pan":"4447","exp_year":2024,"exp_month":5,"holder":"John Smith"},"authorization_information":{"url":"http:\/\/...","data":"some_data"}}';
+        $json = '{"id":"foo","amount":"55.00","type":"bar","payment_method":"card","payment_instrument":{"card_brand":"Visa","pan":"4447","exp_year":'.(date('Y')+4).',"exp_month":11,"holder":"James Bond"},"authorization_information":{"url":"http:\/\/...","data":"some_data"}}';
 
         $payment = new Payment\Payment();
         $payment->unserialize($json);
@@ -50,7 +40,7 @@ class PaymentTest extends ClientTestCase
         $this->assertInstanceOf('Cardinity\Method\Payment\AuthorizationInformation', $payment->getAuthorizationInformation());
         $this->assertSame('http://...', $payment->getAuthorizationInformation()->getUrl());
         $this->assertInstanceOf('Cardinity\Method\Payment\PaymentInstrumentCard', $payment->getPaymentInstrument());
-        $this->assertSame('John Smith', $payment->getPaymentInstrument()->getHolder());
+        $this->assertSame('James Bond', $payment->getPaymentInstrument()->getHolder());
     }
 
     /**