+++ tests/ThreeDS2Test.php
@@ -22,19 +22,65 @@ class ThreeDS2Test extends ClientTestCase
         foreach ($browserInfo as $k => $v) $this->assertEquals($v, $testBrowserInfo[$k]);
     }
 
+    /**
+     * @dataProvider optionalBrowserInfoProvider
+     */
+    public function testBrowserInfoOptionalParams($args)
+    {
+        $browserInfo = $this->getBrowserInfo($args);
+        $testBrowserInfo = $this->browserInfoProvider($args);
+        foreach ($browserInfo as $k => $v) $this->assertEquals($v, $testBrowserInfo[$k]);
+    }
+
+    public function optionalBrowserInfoProvider()
+    {
+        return [
+            [['ip_address' => '0.0.0.0']],
+            [['javascript_enabled' => false]],
+            [['javascript_enabled' => true]],
+            [['java_enabled' => false]],
+            [['java_enabled' => true]],
+            [['ip_address' => '0.0.0.0', 'javascript_enabled' => true]],
+            [['ip_address' => '0.0.0.0', 'java_enabled' => true]],
+            [['ip_address' => '0.0.0.0', 'java_enabled' => false, 'javascript_enabled' => true]],
+            [['java_enabled' => false, 'javascript_enabled' => true]],
+        ];
+    }
+
     public function testAddressObject()
     {
         $address = $this->getAddress();
         $this->assertInstanceOf('Cardinity\Method\Payment\Address', $address);
     }
 
-    public function testAddressObjectsParams()
+    public function testAddressObjectParams()
     {
         $address = $this->getAddress();
         $testAddr = $this->addressProvider();
         foreach ($address as $k => $v) $this->assertEquals($v, $testAddr[$k]);
     }
 
+    /**
+     * @dataProvider optionalAddressProvider
+     */
+    public function testAddressObjectOptionalParams($args)
+    {
+        $address = $this->getAddress($args);
+        $testAddr = $this->addressProvider($args);
+        // print_r($testAddr);exit;
+        foreach ($address as $k => $v) $this->assertEquals($v, $testAddr[$k]);
+    }
+
+
+    public function optionalAddressProvider()
+    {
+        return [
+            [['address_line2' => 'address line 2']],
+            [['address_line3' => 'address line 3']],
+            [['state' => 'vilnius']],
+            [['address_line2' => 'address line 2', 'state' => 'vilnius']],
+        ];
+    }
     public function testThreeDS2DataObject()
     {
         $treeDS2Data = $this->getThreeDS2DataMandatoryData();
@@ -94,35 +140,43 @@ class ThreeDS2Test extends ClientTestCase
         return $paymentArr;
     }
     
-    public function ThreeDS2DataProvider()
+    public function ThreeDS2DataProvider(array $args = [])
     {
-        return [
-            "notificationUrl"=>"https:\/\/notification.url\/",
-            "browserInfo"=> $this->browserInfoProvider()
+        $data = [
+            "notification_url"=>"https:\/\/notification.url\/",
+            "browser_info"=> $this->browserInfoProvider()
         ];
+        if ($args) array_push($data, $args);
+
+        return $data;
     }
 
-    public function addressProvider()
+    public function addressProvider(array $args = [])
     {
-        return [
+        $address = [
             "address_line1"=>"adress 1",
             "city"=>"city",
             "country"=>"LT",
             "postal_code"=>"02245"
         ];
+        if ($args) foreach ($args as $k => $v) $address[$k] = $v;
+
+        return $address;
     }
 
-    public function browserInfoProvider()
+    public function browserInfoProvider(array $args = [])
     {
-        return [
+        $info = [
             "accept_header"=>"HTTP accept header.",
             "browser_language"=>"cardholder language IETF BCP 47.",
             "screen_width"=>600,
             "screen_height"=>400,
             "challenge_window_size"=>"600x400",
             "user_agent"=>"agent James Bond",
-            "color_depth"=>123,
+            "color_depth"=>24,
             "time_zone"=>-60
         ];
+        if ($args) foreach ($args as $k => $v) $info[$k] = $v;
+        return $info;
     }
 }